backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
python-multipart
python-docx
pdfplumber